Design: add line border when click events occurs Rename: src/Window -> src/View
