+++ src/View/PaintBoardApplication.java
@@ -1,6 +1,5 @@
-package Window;
+package View;
 
-import javax.swing.*;
 import java.awt.*;
 
 public class PaintBoardApplication {
