+++ src/View/ToolBar.java
@@ -1,13 +1,11 @@
-package Window;
+package View;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.Button;
 
 public class ToolBar extends JToolBar {
         private static final String[] iconList = {"backgroundColor", "brush", "deleteAll", "group", "ungroup", "open", "lineCol", "lineEraser", "paste", "save", "pixelEraser", "save", "shapeColor", "textBox"};
-        private ImageIcon[] orgIcon;
-        private ImageIcon[] chgIcon;
+
 
         public ToolBar() {
                 setPreferredSize(new Dimension(200, 400));
